update typings based on the structural rule
