+++ src/py/rizemind/strategies/contribution/sampling/typing.py
@@ -1,0 +1,8 @@
+from typing import Protocol
+
+
+class RandomDeterministicInterface(Protocol):
+    def get_number_of_sets(self, round_id: int) -> int: ...
+    def get_number_of_participants(self, round_id: int) -> int: ...
+    def get_nth_participant(self, round_id: int, n: int) -> str: ...
+    def get_nth_set(self, round_id: int, n: int) -> str: ...
