+++ src/py/rizemind/strategies/typing.py
@@ -1,0 +1,11 @@
+from typing import Protocol
+
+from eth_typing import ChecksumAddress
+from flwr.common import FitRes
+
+
+class SupportsFitResStore(Protocol):
+    def insert(self, client_id: ChecksumAddress, fit_res: FitRes) -> None: ...
+    def get(self, client_id: ChecksumAddress) -> FitRes: ...
+    def clear(self) -> None: ...
+    def items(self) -> list[tuple[ChecksumAddress, FitRes]]: ...
